RFC: fix codeclimate warnings
codeclimate noted a few similiar code chunks and it looks
like it is worth fixing them.
